backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic>=2.7.4,<3.0.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-dotenv==1.0.0
neo4j==5.14.1
pytest==7.4.0
langgraph
langchain-google-genai==3.0.3
pydantic[email]>=2.7.4,<3.0.0
bcrypt==4.0.1
pandas==2.1.4
google-generativeai==0.3.2
